feat(User): create permissions edit page

+++ backend/src/resolvers/Mutation.js
@@ -2,6 +2,7 @@ const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const {randomBytes} = require('crypto');
 const {promisify} = require('util');
+const {hasPermission} = require("../utils");
 
 const mutations = {
 	async createItem(parent, args, ctx, info) {
@@ -32,7 +33,18 @@ const mutations = {
 	},
 	async deleteItem(parent, args, ctx, info) {
 		const where = {id: args.id};
-		const item = await ctx.db.query.item({where}, `{ id title}`);
+		const item = await ctx.db.query.item({where}, `{ id title user { id }}`);
+		const ownsItem = item.user.id === ctx.request.userId;
+		const hasPermissions = ctx.request.user.permissions.some(permission => {
+				const isAdmin = permission === 'ADMIN';
+				const hasDeletePermission = permission === 'ITEMDELETE';
+				return isAdmin || hasDeletePermission;
+			}
+		);
+
+		if (!ownsItem && !hasPermissions) {
+			throw new Error("You don't have permission to delete this item!");
+		}
 		return ctx.db.mutation.deleteItem({where}, info);
 	},
 	async signup(parent, args, ctx, info) {
@@ -42,7 +54,7 @@ const mutations = {
 			data: {
 				...args,
 				password,
-				permissions: {set: ['USER']}
+				permissions: {set: ['USER', 'PERMISSIONUPDATE']}
 			}
 		}, info);
 		const token = jwt.sign({userId: user.id}, process.env.APP_SECRET);
@@ -121,6 +133,30 @@ const mutations = {
 		});
 		return updatedUser;
 	},
+	async updateUserPermissions(parent, args, ctx, info) {
+		const userId = ctx.request.userId;
+		if (!userId) {
+			throw 'You are not logged in';
+		}
+		const currentUser = await ctx.db.query.user({
+			where: {
+				id: userId
+			}
+		}, info);
+		if (currentUser) {
+			hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE'])
+			return ctx.db.mutation.updateUser({
+				data: {
+					permissions: {
+						set: args.permissions
+					}
+				},
+				where: {
+					id: args.userId
+				}
+			}, info)
+		}
+	}
 };
 
 module.exports = mutations;
